jicho/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
apscheduler==3.10.4
anthropic==0.29.0
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==23.2.1
pydantic==2.7.4
